requirements.txt
fastapi
uvicorn
openai
PyMuPDF
streamlit
requests
python-dotenv
sentence-transformers
faiss-cpu
python-multipart
python-dotenv
logging